requirements.txt
neo4j
mcp
streamlit
google-generativeai
